Move layer identity to schemas/
